Remove use of sys.atexit to call main funciton
As functions registered with sys.atexit are intended be called as the
interpreter is shutting down, various modules and interpreter services
may already be unloaded when the function is called. This makes it a
fragile way to call main functions.

Instead the method of stack inspection is used to determine if a
function is being defined in the __main__ module and executed
appropriately.